+++ lib/bwu_docker_v1x15_to_v1x19.dart
@@ -1,5 +1,0 @@
-library bwu_docker_v1x15_to_v1x19;
-
-export 'src/v1x15_to_v1x19/remote_api.dart';
-export 'src/v1x15_to_v1x19/data_structures.dart';
-