+++ lib/bwu_docker_v1x20.dart
@@ -1,4 +1,0 @@
-library bwu_docker_v1x20;
-
-export 'src/v1x20/remote_api.dart';
-export 'src/v1x20/data_structures.dart';