+++ lib/tasks.dart
@@ -1,5 +1,5 @@
 /// Common higher level Docker tasks built on top of the low level remote API.
-library bwu_docker.tasks;
+library bwu_docker_v1x15_to_v1x19.tasks;
 
-export 'src/v1x20/tasks.dart';
-export 'src/v1x20/data_structures.dart';
+export 'src/v1x15_and_up/tasks.dart';
+export 'src/v1x15_and_up/data_structures.dart';