+++ lib/tasks_v1x20.dart
@@ -1,5 +1,0 @@
-/// Common higher level Docker tasks built on top of the low level remote API.
-library bwu_docker_v1x20.tasks;
-
-export 'src/v1x20/tasks.dart';
-export 'src/v1x20/data_structures.dart';