+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## UNRELEASED
-- Doesn't throw on unexpected JSON items anymore. It can be enabled by setting
+- Some code restructuring because at some point it seemed necessary to split 
+ code int pre 1.20 and 1.20 or higher but actually wasn't. 
+- Doesn't throw on unexpected JSON items anymore. This can be enabled by setting
 `bool doCheckSurplusItems = true`;
-- Support for Docker API v 1.20 # TODO (zoechi) implement new features
+- Partial support for Docker API v 1.20 and v.121 # TODO (zoechi) implement all new features
 - support in `DockerConnection` to use a specific remote API version
 - improve Docker container creation for running tests
 - apply new dartformat